Write image to buffer instead of directly to cout
The rendered image is written to a buffer in memory before it is written to cout. This buffer is then written to cout to remain consistent with the previous behavior. Writing to a buffer is a prerequisite for implementing multithreading on the rows of the image (otherwise, rows would be output in random order).