Multithreading example with OpenMP
This main file is similar to the single-threaded version except that it adds an OpenMP #pragma for multithreaded rendering of the image.

Needs to be compile with -fopenmp. Unfortunately setup of OpenMP on Mac OSX is not straightforward. I've tested the OpenMP parallelization only on Windows (which requires additional changes as drand48() is not available).