Addresses issue: #9
Fixed the materials to use the correct normals for both lambertian and metal, which allows correct scattering for intersection on the inside.
Added a small offset at each intersection along the normal in order to avoid self-intersection (no visible artifacts even with /fp:fast on the test scene).
Added a comment for the equality: schlick(cosine, ref_idx) = schlick(cosine, ni_over_nt)